requirements.txt
fastapi
uvicorn
vercel-kv
pandas
python-dotenv
